Remove contract namespace
