+++ tsconfig.json
@@ -54,6 +54,9 @@
       "@/subscribers/*": [
         "src/subscribers/*"
       ],
+      "@/subscribers/contracts/*": [
+        "src/subscribers/contracts/*"
+      ],
     },
     "target": "es2020",
     "lib": [
